requirements.txt
# HTTP API Framework
fastapi==0.104.1
uvicorn==0.24.0

# System Monitoring (matching Phase 1)
psutil==5.9.6
docker==6.1.3

# HTTP Client
aiohttp==3.9.1
requests==2.31.0

# Configuration & Logging
python-dotenv==1.0.0
PyYAML==6.0.1

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1